feat(trait): support Camel cloud native properties
Beside that, we're deprecating the access to Secret and Configmap by either the CLI or the operator

Closes #5863
